fetch lyrics on song play rather than on request
